+++ packages/pigeon/CHANGELOG.md
@@ -1,9 +1,12 @@
-## 13.2.0
+## 14.0.0
 
+* **Breaking change** [dart] Renames locally defined host API variables.
+  * Host api static field `codec ` changed to `pigeonChannelCodec`.
 * Adds named parameters to host API methods.
 * Adds optional parameters to host API methods.
 * Adds default values for class constructors and host API methods.
 * Adds `isEnum` and `isClass` to `TypeDeclaration`s.
+
 
 ## 13.1.0